feat: request 값 변경

+++ src/main/java/com/partymaker/party/dto/request/Character.java
@@ -1,11 +1,9 @@
 package com.partymaker.party.dto.request;
 
-import java.math.BigInteger;
-
 public record Character(
         String characterName,
         String ownedName,
         boolean isBuffer,
-        BigInteger power
+        long power
 ) {
 }
